alteracoes
