+++ src/main/java/ifpb/edu/br/marketplace/model/Pedido.java
@@ -1,0 +1,4 @@
+package ifpb.edu.br.marketplace.model;
+
+public class Pedido {
+}
