+++ src/main/java/ifpb/edu/br/marketplace/model/Usuario.java
@@ -1,0 +1,4 @@
+package ifpb.edu.br.marketplace.model;
+
+public class Usuario {
+}
