+++ src/main/java/ifpb/edu/br/marketplace/service/NotificacaoService.java
@@ -1,8 +1,8 @@
-package ifpb.edu.br.service;
-
-import ifpb.edu.br.model.Notificacao;
+package ifpb.edu.br.marketplace.service;
 
 import java.util.List;
+
+import ifpb.edu.br.marketplace.model.Notificacao;
 
 interface NotificacaoService {
     void enviarNotificacao(String usuarioId, String mensagem);
