+++ src/main/java/ifpb/edu/br/marketplace/service/PedidoService.java
@@ -1,8 +1,8 @@
-package ifpb.edu.br.service;
-
-import ifpb.edu.br.model.Pedido;
+package ifpb.edu.br.marketplace.service;
 
 import java.util.List;
+
+import ifpb.edu.br.marketplace.model.Pedido;
 
 public interface PedidoService {
     Pedido criarPedido(Pedido pedido);
