--- src/main/java/ifpb/edu/br/model/Categoria.java
@@ -1,4 +1,0 @@
-package ifpb.edu.br.model;
-
-public class Categoria {
-}
