--- src/main/java/ifpb/edu/br/model/Pedido.java
@@ -1,4 +1,0 @@
-package ifpb.edu.br.model;
-
-public class Pedido {
-}
